design: UniversitySearchInput 디자인 수정
