+++ public/svgs/search/index.ts
@@ -1,0 +1,3 @@
+import IconSearch from "./search.svg";
+
+export { IconSearch };
